back/requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.4
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
MarkupSafe==2.1.1
Pillow==9.1.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tinify==1.5.2
urllib3==1.26.9
whitenoise==6.0.0
psycopg2-binary==2.9.1
python-telegram-bot==13.12
